requirements.txt
streamlit
groq
pymupdf
Pillow
python-dotenv